refactor(desktop): extract per-tab observability state mutators
Move the stateful side of browser observability — the listener
callbacks that record into a tab's consoleEntries / errorEntries /
requests / requestOrder — into a new module
browser-pane/tab-observability.ts. The pure formatting layer
(observability.ts) and the tab record's state shape stay where they
were; only the upsert / track / lookup logic moved.

Moved:
- browserTabForWebContentsId
- appendBrowserObservedError
- upsertBrowserRequestRecord
- trackBrowserRequestStart / Headers / Completion / Failure

main.ts adds an `eachBrowserTabRecord()` generator that walks every
workspace + space + agent-session-space, passed as the only dep to
createTabObservability(). Module surface stays small (1 dep).

main.ts -240. Typecheck clean.
